requirements.txt
ffmpeg_python==0.2.0
halo==0.0.31
openai==0.27.6
openai_whisper==20231117
